Added additional exclusions for a handful of services that cause FP's for the Mix Matched Rules ("N-able Technologies", "Microsoft BizTalk Server", and "stealthbits" as well as added no case for all of the contents / content nots.
Added "After" and Thresholding for Cryptomix Malware rule. This is commonly an FP for ".error" and ".cnc" extensions. Ransomware is SUPER loud and will (likely) encrypt more than 10 files in 5 min. Also, if a legitmate incident does occure, to prevent flooding to the SOC I put a "limit" of 5 every 5 min by source.